% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPatchseqTaxonomy.R
\name{addPatchseqQCMetrics}
\alias{addPatchseqQCMetrics}
\title{Save marker genes for patchSeqQC}
\usage{
addPatchseqQCMetrics(
  AIT.anndata,
  mode.name = AIT.anndata$uns$mode,
  subclass.column = "subclass_label",
  class.column = "class_label",
  off.target.types = NULL,
  subclass.subsample = 100,
  num.markers = 50,
  taxonomyDir = file.path(AIT.anndata$uns$taxonomyDir),
  add.dendrogram.markers = TRUE,
  ...
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{mode.name}{A name to identify the taxonomy mode to add QC metrics for (default is the current mode). If the mode does not yet exist, buildPatchseqTaxonomy will call buildTaxonomyMode using the sampled data from the \code{class.column}, \code{subclass.column}, and \code{off.target.types}.}

\item{subclass.column}{Column name corresponding to the moderate-resolution cell types used for the cell types of interest (default = "subclass_label").}

\item{class.column}{Column name corresponding to the low-resolution cell types used for the off-target cell types (default = "class_label").}

\item{off.target.types}{A character vector of off-target (also known as 'contamination') cell types.  This must include at least one of the cell types found in "class.column" and/or "subclass.column" (both columns are checked)}

\item{subclass.subsample}{The number of cells to retain for PatchseqQC contamination calculation (default = 100, probably no need to change).}

\item{num.markers}{The maximum number of markers to calculate per node per direction (default = 50)}

\item{taxonomyDir}{The location to save shiny output (default = current working directory).}

\item{add.dendrogram.markers}{If TRUE (default=TRUE), will also add dendrogram markers to prep the taxonomy for tree mapping. Default is TRUE because the membership values calculated here as well as the associated tree mapping capabilities is required for a subset of QL metrics, including KL divergence calculations.}

\item{...}{Additional variables to be passed to \code{addDendrogramMarkers} or \code{buildTaxonomyMode}

The following variables are added to AIT.anndata$uns:
$dend[\link{mode.name}]
$filter[\link{mode.name}]
$QC_markers[\link{mode.name}]
...$markers,
...$countsQC,
...$cpmQC,
...$classBr,
...$subclassF,
...$allMarkers
$memb[\link{mode.name}]  # Only if add.dendrogram.markers=TRUE
...$memb.ref,
...$map.df.ref}
}
\value{
AIT.anndata An updated AIT.anndata variable with the above content added to AIT.anndata$uns for the relevant mode.name.
}
\description{
This function saves all the variables required for applying the patchseq QC algorithm \code{pathseqtools} (which is an more flexible version of the \code{patchSeqQC} algorithm) to AIT.anndata$uns. This is only used for patch-seq analysis.  Requirements for input include:
}
