% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPatchseqTaxonomy.R
\name{buildPatchseqTaxonomy}
\alias{buildPatchseqTaxonomy}
\title{Save marker genes and QC metrics for patchSeqQC}
\usage{
buildPatchseqTaxonomy(...)
}
\arguments{
\item{...}{Variables to be passed to \code{buildPatchseqTaxonomy}}
}
\value{
AIT.anndata An updated AIT.anndata variable.
}
\description{
This is the same as addPatchseqQCMetrics, but the name buildPatchseqTaxonomy is kept for back-compatibility
}
